+++ packages/credential-sdk/src/types/generic/typed-any.js
@@ -1,5 +1,0 @@
-import createPlaceholder from './create-placeholder';
-
-const Any = (any) => any;
-
-export default createPlaceholder(Any);