+++ packages/credential-sdk/src/types/offchain-signatures/public-keys/id.js
@@ -1,5 +1,10 @@
-import { TypedNumber, TypedUUID } from '../../generic';
+import { TypedNumber, TypedUUID, anyOf } from '../../generic';
 
 export class DockPublicKeyId extends TypedNumber {}
 
 export class CheqdPublicKeyId extends TypedUUID {}
+
+export class DockOrCheqdPublicKeyId extends anyOf(
+  DockPublicKeyId,
+  CheqdPublicKeyId,
+) {}